Removing merge-strategy and moving logic into the token fetcher
