+++ packages/token-sdk/src/metadata/index.ts
@@ -1,4 +1,3 @@
 export * from "./coingecko-provider";
 export * from "./metaplex-provider";
-export * from "./merge-strategy";
 export * from "./models";
